+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Now, instead, the "Executor" classes implement the `getBinary()` method, which r
   should return the names of the binaries (as a string or array of strings, e.g. in the case of `MysqlExecutor`) related
   to the respective driver; The `DATABASE_BACKUP_EXECUTABLES` constant, previously defined in the bootstrap, has been
   removed;
+* the `AbstractBackupUtility::getProcess()` method has been removed;
 * now the constructor of `AbstractExecutor` also requires the `$OperationType` argument;
 * the `AbstractBackupUtility::_getDriver()` had been deprecated and has now been removed;
 * the `AbstractExecutor::__call()` magic method, which provided backwards compatibility for deprecated